Crate copyless

Helper extensions of standard containers that allow memcopy-less operation.

Structs

BoxAllocation

A typesafe helper that stores the allocated pointer without the data initialized.

VecAllocation

A typesafe helper that separates new value construction from vector growing, allowing LLVM to ideally construct the element in place.

Enums

VecEntry

An entry into a vector, similar to std::collections::hash_map::Entry.

Traits

BoxHelper

Helper trait for a Box type that allocates up-front.

VecHelper

Helper trait for a Vec type that allocates up-front.